5º período/TI5/TI5/back/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
influxdb-client==1.46.0